refactor: move site files
