+++ src/pages/index.tsx
@@ -1,8 +1,8 @@
 import { type NextPage } from "next";
 import { useSession } from "next-auth/react";
-import LogoutScreen from "../comp/LogoutScreen";
-import SiteWrapper from "../comp/SiteWrapper";
 import Link from "next/link";
+import SiteWrapper from "../comp/SiteWrapper";
+import LogoutScreen from "../sites/LogoutScreen";
 import { hatIcon, listIcon, sparklesIcon } from "../utils/icons";
 
 const Home: NextPage = () => {
